LayoutTests/media/media-source/media-source-append-buffer-with-append-window-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
EXPECTED (sourceBuffer.appendWindowStart == '0') OK
EXPECTED (sourceBuffer.appendWindowEnd == 'Infinity') OK
RUN(sourceBuffer.appendWindowEnd = 7.5)
RUN(sourceBuffer.appendWindowStart = 3.5)
RUN(sourceBuffer.appendBuffer(mediaSegment))
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.start(0) == '4') OK
EXPECTED (sourceBuffer.buffered.end(0) < '7.5') OK
RUN(sourceBuffer.appendWindowStart = 0)
RUN(sourceBuffer.appendBuffer(mediaSegment))
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '1') OK
EXPECTED (sourceBuffer.buffered.start(0) == '0') OK
EXPECTED (sourceBuffer.buffered.end(0) < '7.5') OK
RUN(sourceBuffer.appendWindowEnd = 14.5)
RUN(sourceBuffer.appendWindowStart = 11)
RUN(sourceBuffer.appendBuffer(mediaSegment))
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '2') OK
EXPECTED (sourceBuffer.buffered.start(1) == '12') OK
EXPECTED (sourceBuffer.buffered.end(1) < '14.5') OK
END OF TEST
